wpt_babel/content-security-policy/navigation/to-javascript-parent-initiated-parent-csp.html

<!DOCTYPE html>
<head>
 <meta content="script-src 'self' 'nonce-abc'" http-equiv="content-security-policy"/>
 <script src="/resources/testharness.js">
 </script>
 <script src="/resources/testharnessreport.js">
 </script>
</head>
<body>
 <iframe src="support/frame-with-csp.sub.html">
 </iframe>
 <div id="special_div" onclick="frames[0].location.href = 'javascript:parent.postMessage(\'executed\', \'*\')'">
 </div>
 <script nonce="abc">
  "use strict";

var t = async_test("Should not have executed the javascript url");
frames[0].addEventListener('load', function () {
  window.onmessage = t.step_func(function (e) {
    if (e.data == "executed") assert_true(false, "Javascript url executed");
  });
  window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
    assert_equals(e.blockedURI, 'inline');
    assert_equals(e.violatedDirective, 'script-src-attr');
  }));
  document.getElementById('special_div').click();
});
 </script>
</body>
